Move document parsing to Meeseeks.Parser and update
  - Move Document.new to Parser.parse_tuple_tree
  - Update parse_tuple_tree to add roots and children in first-to-last order (deoptimization)
  - Remove reversal of children in Document.children and Document.descendants
  - Add test to ensure identical output when parsing a string and parsing a tuple tree
