+++ lib/meeseeks/document.ex
@@ -24,7 +24,7 @@ defmodule Meeseeks.Document do
                             {"p", [], ["3"]}]}]}]}
   {...}
 
-  iex> document = Meeseeks.Document(tuple_tree)
+  iex> document = Meeseeks.Parser.parse(tuple_tree)
   %Meeseeks.Document{
     id_counter: 12,
     roots: [1],
@@ -58,8 +58,7 @@ defmodule Meeseeks.Document do
   """
 
   alias Meeseeks.Document
-  alias Meeseeks.Document.{Comment, Data, Doctype, Element, Node, Text}
-  alias Meeseeks.TupleTree
+  alias Meeseeks.Document.{Element, Node}
 
   defstruct id_counter: nil, roots: [], nodes: %{}
 
@@ -68,133 +67,6 @@ defmodule Meeseeks.Document do
   @type t :: %Document{id_counter: node_id | nil,
                        roots: [node_id],
                        nodes: %{optional(node_id) => node_t}}
-
-  # Build
-
-  @doc """
-  Creates a document from a `Meeseeks.TupleTree`.
-
-  Indexes nodes in depth-first order.
-
-  Generally be called via `Meeseeks.Parser.parse`, not directly.
-  """
-  @spec new(TupleTree.t) :: Document.t
-
-  def new(tuple_tree) when is_list(tuple_tree) do
-    add_root_nodes(%Document{}, tuple_tree)
-  end
-
-  def new(tuple_tree) when is_tuple(tuple_tree) do
-    add_root_node(%Document{}, tuple_tree)
-  end
-
-  defp add_root_nodes(document, roots) do
-    Enum.reduce(roots, document, &(add_root_node &2, &1))
-  end
-
-  defp add_root_node(document, {tag, attributes, children}) do
-    id = next_id(document.id_counter)
-    [ns, tg] = split_namespace_from_tag(tag)
-    node = %Element{id: id,
-                    namespace: ns,
-                    tag: tg,
-                    attributes: attributes}
-    %{document |
-      id_counter: id,
-      roots: [id | document.roots],
-      nodes: insert_node(document.nodes, node)}
-    |> add_child_nodes(id, children)
-  end
-
-  defp add_root_node(document, {:comment, comment}) do
-    id = next_id(document.id_counter)
-    node = %Comment{id: id, content: comment}
-    %{document |
-      id_counter: id,
-      roots: [id | document.roots],
-      nodes: insert_node(document.nodes, node)}
-  end
-
-  defp add_root_node(document, {:doctype, type, public, system}) do
-    id = next_id(document.id_counter)
-    node = %Doctype{id: id, type: type, public: public, system: system}
-    %{document |
-      id_counter: id,
-      roots: [id | document.roots],
-      nodes: insert_node(document.nodes, node)}
-  end
-
-  defp add_root_node(document, _other) do
-    document
-  end
-
-  defp add_child_nodes(document, parent_id, children) do
-    Enum.reduce(children, document, &(add_child_node &2, parent_id, &1))
-  end
-
-  defp add_child_node(document, parent, {tag, attributes, children}) do
-    id = next_id(document.id_counter)
-    [ns, tg] = split_namespace_from_tag(tag)
-    node = %Element{parent: parent,
-                    id: id,
-                    namespace: ns,
-                    tag: tg,
-                    attributes: attributes}
-    %{document |
-      id_counter: id,
-      nodes: insert_node(document.nodes, node)}
-    |> add_child_nodes(id, children)
-  end
-
-  defp add_child_node(document, parent, {:comment, comment}) do
-    id = next_id(document.id_counter)
-    node = %Comment{parent: parent, id: id, content: comment}
-    %{document |
-      id_counter: id,
-      nodes: insert_node(document.nodes, node)}
-  end
-
-  defp add_child_node(document, parent, text) when is_binary(text) do
-    id = next_id(document.id_counter)
-    parent_node = get_node(document, parent)
-    if parent_node.tag == "script" or parent_node.tag == "style" do
-      node = %Data{parent: parent, id: id, content: text}
-      %{document |
-        id_counter: id,
-        nodes: insert_node(document.nodes, node)}
-    else
-      node = %Text{parent: parent, id: id, content: text}
-      %{document |
-        id_counter: id,
-        nodes: insert_node(document.nodes, node)}
-    end
-  end
-
-  defp add_child_node(document, _parent, _other) do
-    document
-  end
-
-  defp next_id(nil), do: 1
-  defp next_id(n), do: n + 1
-
-  defp split_namespace_from_tag(maybe_namespaced_tag) do
-    case :binary.split(maybe_namespaced_tag, ":", []) do
-      [tg] -> [nil, tg]
-      [ns, tg] -> [ns, tg]
-    end
-  end
-
-  defp insert_node(nodes, %{parent: nil, id: id} = node) do
-    Map.put(nodes, id, node)
-  end
-
-  defp insert_node(nodes, %{parent: parent, id: child} = node) do
-    parent_node = Map.get(nodes, parent)
-    children = parent_node.children
-    nodes
-    |> Map.put(child, node)
-    |> Map.put(parent, %{parent_node | children: [child | children]})
-  end
 
   # Query
 
@@ -220,8 +92,7 @@ defmodule Meeseeks.Document do
   @spec children(Document.t, node_id) :: [node_id]
   def children(document, node_id) do
     case get_node(document, node_id) do
-      %Document.Element{children: children} ->
-        Enum.reverse(children)
+      %Document.Element{children: children} -> children
       _ -> []
     end
   end
@@ -236,8 +107,7 @@ defmodule Meeseeks.Document do
   @spec descendants(Document.t, node_id) :: [node_id]
   def descendants(document, node_id) do
     case get_node(document, node_id) do
-      %Element{children: cs} ->
-        children = Enum.reverse(cs)
+      %Element{children: children} ->
         Enum.flat_map(children, &[&1 | descendants(document, &1)])
       _ -> []
     end
